Finalizando partida de treino

+++ lib/screens/PreMatch.dart
@@ -9,12 +9,14 @@ import 'package:flutter/material.dart';
 import 'package:jogodavelha/constants/Colors.dart';
 import 'package:jogodavelha/constants/Messages.dart';
 import 'package:jogodavelha/screens/GameMultiplayer.dart';
+import 'package:jogodavelha/screens/Game.dart';
 import 'package:jogodavelha/screens/Lobby.dart';
 import 'package:jogodavelha/storage/CurrentUser.dart';
 import '../models/LobbyModel.dart';
 import '../models/User.dart';
 import '../models/Match.dart';
 import '../services/Api.dart';
+import '../storage/Bot.dart';
 
 LobbyModel currentLobby;
 
@@ -41,15 +43,21 @@ class _PreMatchState extends State<PreMatch> {
 
   setVariables() async{
     startTimer();
-    if(currentLobby.player1 == CurrentUser.user.id){
+    if(currentLobby.player1 == CurrentUser.user.id) { //Se eu sou o player 1, economizo uma requisição
       _player1 = CurrentUser.user;
-      Api.getUser(currentLobby.player2).then((p2) {
+      if(currentLobby.player2 == Bot.botInfos.id){ //Se sim, é um treino. Não preciso pegar usuário no BD -> Bot nunca será player1
         setState(() {
-          _player2 = p2;
+          _player2 = Bot.botInfos;
         });
-      });
-    }else{
-      _player2 = CurrentUser.user;
+      }else{ //Não é um treino
+        Api.getUser(currentLobby.player2).then((p2) {
+          setState(() {
+            _player2 = p2;
+          });
+        });
+      }
+    } else{
+      _player2 = CurrentUser.user; //Se não sou player 1, sou o player 2, economizo uma requisição
       Api.getUser(currentLobby.player1).then((p1) {
         setState(() {
           _player1 = p1;
@@ -66,14 +74,20 @@ class _PreMatchState extends State<PreMatch> {
       currentMatch.matchtoken = currentLobby.token;
       currentMatch.playerOfTheRound = _player1.id;
       currentMatch.timestamp = DateTime.now().millisecondsSinceEpoch.toString();
-      await Api.updateMatch(currentMatch);
-      if(currentLobby.player1 == CurrentUser.user.id){ //Só um pode apagar o Lobby
-        await Api.deleteLobby(currentLobby);
+
+      if(_player2.id != Bot.botInfos.id){ //Não é treino
+        await Api.updateMatch(currentMatch);
+        if(currentLobby.player1 == CurrentUser.user.id){ //Só um pode apagar o Lobby
+          await Api.deleteLobby(currentLobby);
+        }
+        Navigator.pushAndRemoveUntil(context,
+            MaterialPageRoute(builder: (BuildContext context) => GameMultiplayer(currentMatch, _player1, _player2)), (Route<dynamic> route) => false);
+      }else{ // É um treino, não precisa salvar
+        Navigator.pushAndRemoveUntil(context,
+            MaterialPageRoute(builder: (BuildContext context) => Game(currentMatch, _player1, _player2)), (Route<dynamic> route) => false);
       }
-      Navigator.pushAndRemoveUntil(context,
-          MaterialPageRoute(builder: (BuildContext context) => GameMultiplayer(currentMatch, _player1, _player2)), (Route<dynamic> route) => false);
     }catch(e){
-      print(e);
+      print(e); //Todo: Tratar Exception com modal
     }
   }
 
@@ -82,7 +96,7 @@ class _PreMatchState extends State<PreMatch> {
     Timer _timer = new Timer.periodic(
       oneSec,
           (Timer timer) {
-        if (_counter == 1) {
+        if (_counter <= 1) {
           setState(() {
             timer.cancel();
             createMatch();
